requirements.txt
aiogram==3.7.0
python-dotenv==1.0.1
aiofiles~=23.2.1
aiohttp==3.9.5
loguru==0.7.2
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-mock==3.12.0